Debug: Fix incorrect handling of STM parameter
The STM parameter (register 3) was incorrect. Use ex_insn instead of
wb_insn to trigger the output. The following code can lead to
incorrect STM events when using wb_insn:

  l.ori r3, r0, 1
  l.nop 0x23
  l.lwz r3, -4(r2)

Instead of emitting (0x23, 1) it can output (0x23, <undefined>). The
problem is that while wb_freeze is still high, r3 is already updated
with new values (or undefined). When wb_freeze goes down, the stored
value got replaced with an undefined value.

By using ex_insn and ex_freeze, the problem is solved. The problem is
that the information is now imprecise relative to ITM events, that is
for the moment okay.