Prototyping: Add DDR memory for ZTEX boards
Add the DDR memory support to the prototyping system. Enable DDR
memory. After a lot of trouble with regards to this, we found we could
simply run the memory at a slower clock speed (200 MHz at the moment),
which works perfectly fine.

Also add a simple application that tests the memory, that writes data
to all memory addresses and reads back.