GUI: Add system id of 2x2 ztex system
This is a temporary workaround until the XML-based system configuration
is done. Since no memory initialisation is possible through the GUI
in a distributed memory system, it only works for the simulation where
the software is loaded from a VMEM file, or if the system was already
initialized through optimsoc_cli.