liboptimsochost: Fix possible crash on disconnect
For some reason the receive thread does not always terminate
properly when calling pthread_cancel/pthread_join, resulting
in a crash when accessing the usb context that's already deleted.
As a fix we insert a explicit cancellation point after each USB
read transfer, which allows the thread to cleanly cancel and
exit.