GUI: Rework to a task-based interface
This commit contains a major rework of the user interface
and the concept behind it. The two main changes are:

- The UI is now task based, similar to Eclipse. Those tasks are
  grouped in tabs, each tab shows a single task and is called "view".
  Think of a view as a workspace or a desk, on which different
  components can be placed depending on the which task is currently
  being worked at.

  Those views are currently static, but are planned to be
  user-definable in the future.

- A first step towards the XML system definitions has been made.

  The XML files describe a system (identified by the System Identifier
  specified in the hardware design, i.e. in the Verilog sources).
  It contains information about the tiles as well as the NoC. This
  information can then be used by the GUI to inform the user about
  the system he/she has in front of him/her, and to make the interaction
  with the system as easy as possible.

  Apart from a hierarchical list of components contained in the system
  a graphical representation is shown in the UI. It turns out to be
  rather hard to represent any possible system configuration in a graphic
  by generating it from a interconnect structure (the NoC description)
  and the tile descriptions. To circumvent this problem, the current
  design uses a manually generated SVG file to represent the system.
  The individual SVG elements can be given a optimsoc:idref attribute
  which refers to a id attribute in the system description XML.
  Through this connection the GUI is able to react to mouse clicks
  on elements in the system view (the SVG file) and let the user
  interact with the system (i.e. initialize a memory).

  The short-term goal is to auto-generate those SVG files automatically
  for simple interconnect structures, i.e. a regular mesh.

Only visible for developers is a change to the SystemInterface, the
former HardwareInterface. Calling methods of this class is now safer,
since they can be called like regular methods, and all liboptimsochost
library calls expected to take longer are automatically dispatched
into a worker thread.

A last major change is a change of license. Since we incorporated
the QCustomPlot library the whole application is now licensed under
GPLv3+. To retain more flexibility for us and other contributors in the
use of the code, all our own code was relicensed to the MIT license.
Note that this does *not* mean that the whole application is now MIT
licensed, unless you replace all GPL code with something else.