Implement OpTiMSoC XML System Descriptions
A OpTiMSoC XML System Description is a XML file which contains a description
of a single OpTiMSoC system. Each system is described by a unique identifier,
the System ID. This ID is set at hardware design time and can be read through
the Debug infrastructure on the host PC. By using the XML description for the
ID the user can be presented with more information about the system, including
the tile and interconnect configuration, the memory sizes, etc.

To make editing of those files as easy as possible, they come in two forms:
a generic form and a simplified/convenience form. The generic form describes
the interconnect network in terms of connections between routers without
any further information about how the network looks (a mesh, a ring, etc.).
Using this description form is quite verbose and hard to edit by hand. That's
why we introduced a convenience form, which allows easy description of
well-known network types, such as a mesh. These convenience forms can then
be transformed into the generic form using a XSLT stylesheet.

Apart from a description of the components, the XML description also contains
a graphic (in SVG format) depicting the system. This can be used for
documentation or for easier interaction of the user with the system. A usage
example is given in the GUI.

To validate this XML description a XML Schema is given.