GUI: Implement improved Sysdesc XML
With the improved System Description XML, the convenience forms
are automatically converted into the generic form and the SVG
representations are also automatically generated.

Unfortunately the XSLT implementation of Qt is not as powerful
as one could wish. For that reason we fall back to libxslt and
need to introduce this as a new library dependency.