SYN: Use environment variable for synplify cmd
Unfortunately Synplify needs to be called with a different name depending
on the installed licenses. This might be synplify, synplify_premier or
synplify_premier_dp.

We work around this limitation by introducing a new environment variable,
called SYNPLIFY, which is used inside the Makefile. This variable contains
the path or the name of the synplify binary to call.

Example:
If you have a Synplify Premier license with Design Planner do the
following before calling make:
export SYNPLIFY=synplify_premier_dp

The default command is still "synplify", so nothing should change if you
don't set this variable.