SystemSW/runtime: Virtual memory refactor
We change the page table representation to reflect the standard
tables. Although this is at least 9kB for each task, this eases the
transition to hardware-based TLB refill.

Clean namespace and allow to create tables from the outside. Comment
old code, there is still some stuff for TLB management we need later..