rtl/na: add compute tile list
Until now we required the sysconfig.c to be provided with each
platform. Most information has already been moved from there to the
network adapter configuration module.

Finally, this moves the last information to the network adapter, that
is the list of compute tiles.

There is a 32 bit register that holds the number of compute tiles and
a list of 16 bit registers that contain the tile identifiers of the
compute tiles.