SystemSW/baremetal: Enable mp-simple
The MP Simple implementation now has a bootstrapping feature: A
network adapter must be enabled once the interrupt handler is
installed and the software can process incoming requests. Before the
receiver mp_simple is not ready to receive messages, only a special
control request is allows to be sent. The receiving mp_simple
automatically answers this and the sending mp_simple _must_ be enabled
to receive this reply.

It is mandatory to query the domain via optimsoc_mp_simple_ctready and
enable the own network adapter with optimsoc_mp_simple_enable before
starting any message passing operation.