gzll: Fix node and task naming
A node is the physical instance running in a compute tile (on a rank).
A task is the node as seen from the application software.

Task IDs are unique in an application, which can span multiple tiles
without knowing. A tupel of <rank,node> is also unique.

Finally, in the future a node can mode to another node. While it is
still identified under the same task id to the application, the node
id changes (<rank,node> -> <rank',node'>).