Get the OpTiMSoC version number from the src tree
We introduce a new helper script, get-version.sh that returns the
version number associated with the code in the source code directory. It
tries to give a meaningful number that enables recreating the same
source tree state later.

We then use this version number as default in the install.py script.