Set prefix of tools to include $OPTIMSOC
Set the hard-coded prefix path inside the installed to a path relative
to the $OPTIMSOC variable. This variable is not read from the
environment, but must be explicitly passed to pkg-config by using the
parameter

pkg-config --define-variable=OPTIMSOC=your/path/to/optismoc

This change brings us one step further to fully relocatable builds. In a
further commit passing this parameter to pkg-config will be automated.