New config structure
We now use a struct as static configuration parameter that contains
all configuration settings. This eases passing those parameters
through the hierarchy and thereby avoids mistakes and increases
readability.

There are two struct: The base_config_t struct contains all
configuration settings. It is used to setup the system. The config_t
struct contains some extra settings, which are derived from the base
settings. You can call the function derive_config to convert a
base_config_t struct into a config_t.

When you add a base variable you need to add it at three positions:

 - In base_config_t

 - In config_t

 - The copy operation in derive_config

If it is a derived setting, you only need to add it to the latter two.

While the use of the struct works pretty well with verilator and
vivado so far, the lack of support for variable-sized arrays in SV
needs us to workaround with "max size" arrays for now. Actually,
classes as parameters would be great..