systemtests: Introduce and use an output filter
Before this commit we always compared the output with the golden
reference directly. This commit introduces a way to define a filter
function which is applied to the golden reference and the test output
before comparing the two files. It can be used to remove variable
content from the output and thus make the tests more robust.

This commit also introduces a filter function which removes the
timestamp from the OpTiMSoC log files.

Note: it's important to explicitly disable the filecmp shallow
comparison feature, as it creates wrong results (changes are not
recognized).