Add API documentation
The API documentation is kind of strange, because it uses doxygen, but
doesn't generate the default html or latex. Instead we generate xml
from it and convert it to rst using a script and the breathe extension
for Sphinx.

From the rst everything can be generated, and this flow has the
advantage we can easily link between reference manual and the API
documentation.