LISNoC: Fixes and deprecate upstream fetching
LISNoC will in the next months be deprecated and become an integral
part of OpTiMSoC. Remove from update script.

Beside that this begins removing all references to the first flag,
which is not necessary. In future only the last flag will be
kept. Beside that the type field will be deprecated as long as
multicast or priorities find proper use in OpTiMSoC.