doc: Automatically handle versions
Instead of keeping track of individual versions the sphinx build can
take care of the actual version.

This uses the 'version' and a 'lastversion' variable in the
documentation. They are constructed from git. It also generates a note
in the installation part that refers to the last stable release when
the documentation is generated in a development branch.