doc: Temporarily hotfix breathe for Sphinx 1.6
In the future we will use an image converter extension for
sphinx. Those are only available with Sphinx 1.6, but breathe fails
with Sphinx>=1.6. Before we pinned Sphinx to 1.5, now we need to
hotfix the issue with breathe and Sphinx 1.6 until the following issue
is merged: https://github.com/michaeljones/breathe/pull/334