Check for tool versions during build
We quite often require very recent versions of Verilator, FuseSoC, or
(less often) of Vivado. Check during the build process if the documented
minimum requirements are installed, and fail the build otherwise. This
should prevent a series of unhappy users trying to figure out bugs that
we already know about.