SRAM: Make naming of parameters and addresses clear
The current implementation regularly confuses byte and word addresses,
as they use the same signal name "addr". This commit fixes these
occurrences, and makes it harder to confuse the two in the future by
explicitly using "waddr" for word addresses and "addr" for byte
addresses. Additionally, MEM_SIZE is renamed to MEM_SIZE_BYTE to give an
indication of the unit.