Make DI max. packet length a parameter
The maximum number of flits in a DI packet is an important parameter, as
it defines the size of the output FIFOs and a couple of other FIFOs in
the system. Previously it was not set in one central place but ignored
or set to differing values across the hierarchy.