Change compute tile listing in NA
So far we maintained a list with IDs of compute tiles, which is very
inefficient. Instead we have a bitmask now of all tiles with the
compute tiles having a bit set at their position. They are grouped in
32 bit groups, and 32 groups are there, which leads to a total of 1024
tiles supported.

The software is adopted accordingly. As this procedure is rather slow,
it is moved to optimsoc_init to fill the same data structure we had
before.