Make mor1kx configuration more customizable
Currently OpTiMSoC does not expose many features of the mor1kx CPU core
through its configuration structure. Especially enabling the FPU and the
performance counter functionality can sometimes be necessary.

This patch does not change the default configuration, but only exposes
the FPU and Performance Counter features.

The debug unit feature is not exposed as config option, as we need to
couple this feature with the presence (or absence) of the CDM module,
which is currently being developed in OSD.