Add Eclipse Oomph installation description
The Eclipse Installer is based on Oomph and can be used to install
customized Eclipse installations. We use this for OpTiMSoC to give users
a easy way of installing Eclipse with all required settings.

Currently, the Easy Installation mode only allows the selection of a
product, not a project. The project part should therefore be considered
experimental and possibly disabled.