Lint: No range selects on integers
Spyglass warns if range selects are used on integers.

---------
The following message appears at the location where range select of a
variable <var-name> of a type <var-type> is used.
[WARNING] Inappropriate range select for <var-type> variable:
"<var-name>"
Where <var-type> can be integer or time.

Potential Issues
A violation is reported when a range select is used for an integer or
time variable.
Consequences of Not Fixing
Range selects are not meaningful for integer variables and time
variables.

How to Debug and Fix
Double-click the violation message. The HDL Viewer window highlights the
line where integer or time variable is used with selected range.
Remove the range select. If you want to select a range from an integer
variable, either mask and shift, or assign the integer to a reg and then
do range select on the reg.
--------