Fixed bug in noc_mux module
The bug would occur when 'activeroute' is zero, 'out_ready' is zero,
and 'in_valid' in non-zero. In this case 'activeroute' would be set but
the 'active' signals would never be updated (because 'req_masked' always
remains zero), causig the module to be stuck to the current channel, until
another last flit arrives on that channel.